changed to work with new slimmer vertex structure.

+++ utils/ic642.c
@@ -2036,7 +2036,7 @@ ic642_make_surface(int max_vertices, int max_faces)
     v->f = (int *)calloc(v->num, sizeof(int)) ;
     if (!v->f)
       ErrorExit(ERROR_NO_MEMORY,"ic642: could not allocate %d faces",v->num);
-    v->n = (int *)calloc(v->num,sizeof(int));
+    v->n = (uchar *)calloc(v->num,sizeof(uchar));
     if (!v->n)
       ErrorExit(ERROR_NO_MEMORY, "ic642: could not allocate %d nbrs", v->n);
     v->num = 0 ;   /* for use as counter in next section */
